+++ CHANGELOG.md
@@ -21,7 +21,10 @@ Adding a requirement of a major version of a dependency is breaking a contract.
 Dropping a requirement of a major version of a dependency is a new contract.
 
 ## [Unreleased]
-[Unreleased]: https://github.com/atlassian-labs/db-replica/compare/release-0.1.25...master
+[Unreleased]: https://github.com/atlassian-labs/db-replica/compare/release-0.1.26...master
+
+## [0.1.26] - 2021-02-18
+[0.1.26]: https://github.com/atlassian-labs/db-replica/compare/release-0.1.25...release-0.1.26
 
 ### Fixed
 - Set connection properties while switching to replica